requirements.txt
asgiref==3.4.1
astroid==2.4.2
autopep8==1.5.3
beautifulsoup4==4.11.1
bs4==0.0.1
bson==0.5.8
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
coverage==5.2.1
dataclasses==0.6
Django==3.2.7
django-cleanup==6.0.0
django-forms-bootstrap==3.1.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
gunicorn==20.1.0
idna==2.10
isort==4.3.21
Jinja2==3.0.1
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
mccabe==0.6.1
Pillow==8.3.2
psycopg2==2.9.9
pycodestyle==2.6.0
python-dateutil==2.8.1
python-dotenv==0.18.0
python-http-client==3.3.2
pytz==2020.1
qrcode==7.1
requests==2.25.1
sentry-sdk==1.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.2.4
starkbank-ecdsa==1.1.1
tld==0.12.6
toml==0.10.1
urllib3==1.26.6
whitenoise==6.0.0
wrapt==1.12.1
